test: clear speech_modal MEDIUM review items
