+++ test/features/speech/ui/widgets/speech_modal/language_dropdown_test.dart
@@ -1,9 +1,6 @@
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:lotti/classes/journal_entities.dart';
-import 'package:lotti/features/journal/model/entry_state.dart';
 import 'package:lotti/features/journal/state/entry_controller.dart';
 import 'package:lotti/features/speech/ui/widgets/speech_modal/language_dropdown.dart';
 import 'package:lotti/get_it.dart';
@@ -13,39 +10,11 @@ import 'package:mocktail/mocktail.dart';
 
 import '../../../../../mocks/mocks.dart';
 import '../../../../../widget_test_utils.dart';
+import 'test_utils.dart';
 
 // ---------------------------------------------------------------------------
 // Fake controller
 // ---------------------------------------------------------------------------
-
-class _FakeEntryController extends EntryController {
-  _FakeEntryController(this._entry);
-
-  final JournalEntity? _entry;
-  final List<String> setLanguageCalls = [];
-
-  @override
-  Future<EntryState?> build({required String id}) {
-    final value = _entry == null
-        ? null
-        : EntryState.saved(
-            entryId: id,
-            entry: _entry,
-            showMap: false,
-            isFocused: false,
-            shouldShowEditorToolBar: false,
-          );
-    if (value != null) {
-      state = AsyncData(value);
-    }
-    return SynchronousFuture(value);
-  }
-
-  @override
-  Future<void> setLanguage(String language) async {
-    setLanguageCalls.add(language);
-  }
-}
 
 // ---------------------------------------------------------------------------
 // Helpers
@@ -85,11 +54,11 @@ JournalEntry _makeJournalEntry() {
   );
 }
 
-Future<_FakeEntryController> _pump(
+Future<FakeEntryController> _pump(
   WidgetTester tester, {
   JournalEntity? entry,
 }) async {
-  final ctrl = _FakeEntryController(entry);
+  final ctrl = FakeEntryController(entry);
   final entryId = entry?.id ?? 'audio-1';
 
   await tester.pumpWidget(
